Moved DOM functionality into DOM.ts
