+++ .gitignore
@@ -197,3 +197,4 @@ FakesAssemblies/
 
 # Visual Studio 6 workspace options file
 *.opt
+*.map
