+++ IEDiagnosticsAdapter.sln
@@ -11,6 +11,9 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Proxy", "Proxy\Proxy.vcxpro
 	EndProjectSection
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "DebuggerCore", "DebuggerCore\DebuggerCore.vcxproj", "{A9505C3C-5EB9-4989-8028-122E8A5F4B46}"
+	ProjectSection(ProjectDependencies) = postProject
+		{1405EE7F-BDA9-4692-B689-200CA8DC1FBF} = {1405EE7F-BDA9-4692-B689-200CA8DC1FBF}
+	EndProjectSection
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Common", "Common\Common.vcxproj", "{DDABCB76-8AAF-454C-86D5-0E13EEA93667}"
 EndProject
