+++ IEDiagnosticsAdapter/resource.h
@@ -6,3 +6,4 @@
 #define IDR_BROWSER_SCRIPT               102
 #define IDR_DEBUGGER_SCRIPT              103
 #define IDR_COMMON_SCRIPT                104
+#define IDR_DOM_SCRIPT                   105
